moving tab js to ts
